+++ src/commands/guild/mod.rs
@@ -1,0 +1,9 @@
+use serenity::framework::standard::macros::group;
+
+mod serverinfo;
+
+use self::serverinfo::*;
+
+#[group]
+#[commands(serverinfo)]
+pub struct Guild;