+++ src/main.rs
@@ -105,6 +105,7 @@ async fn main() -> Result<()> {
         .on_dispatch_error(hooks::dispatch_error)
         .group(&commands::META_GROUP)
         .group(&commands::tags::TAGS_GROUP)
+        .group(&commands::guild::GUILD_GROUP)
         .group(&commands::lastfm::LASTFM_GROUP)
         .group(&commands::users::USERS_GROUP)
         .group(&commands::moderation::MODERATION_GROUP)